Move metrics from flurs.utils.metric to flurs.metrics
